+++ grizzled/file/includer.py
@@ -4,7 +4,7 @@
 Introduction
 ============
 
-The ``grizzled.file.includer`` module contains a class that can be used to
+The `grizzled.file.includer` module contains a class that can be used to
 process includes within a text file, returning a file-like object. It also
 contains some utility functions that permit using include-enabled files in
 other contexts.
@@ -19,31 +19,26 @@ regular expression matches include directives like this::
     %include "/absolute/path/to/file"
     %include "../relative/path/to/file"
     %include "local_reference"
-    %include "http://localhost/path/to/my.config"
 
-Relative and local file references are relative to the including file or URL.
-That, if an ``Includer`` is processing file "/home/bmc/foo.txt" and encounters
+Relative and local file references are relative to the including file. That
+That is, if an `Includer` is processing file "/home/bmc/foo.txt" and encounters
 an attempt to include file "bar.txt", it will assume "bar.txt" is to be found
 in "/home/bmc".
-
-Similarly, if an ``Includer`` is processing URL "http://localhost/bmc/foo.txt"
-and encounters an attempt to include file "bar.txt", it will assume "bar.txt"
-is to be found at "http://localhost/bmc/bar.txt".
 
 Nested includes are permitted; that is, an included file may, itself, include
 other files. The maximum recursion level is configurable and defaults to 100.
 
 The include syntax can be changed by passing a different regular expression to
-the ``Includer`` class constructor.
+the `Includer` class constructor.
 
 Usage
 =====
 
-This module provides an ``Includer`` class, which processes include directives
+This module provides an `Includer` class, which processes include directives
 in a file and behaves like a file-like object. See the class documentation for
 more details.
 
-The module also provides a ``preprocess()`` convenience function that can be
+The module also provides a `preprocess()` convenience function that can be
 used to preprocess a file; it returns the path to the resulting preprocessed
 file.
 
@@ -51,8 +46,6 @@ Examples
 ========
 
 Preprocess a file containing include directives, then read the result:
-
-.. python::
 
     import includer
     import sys
@@ -64,8 +57,6 @@ Preprocess a file containing include directives, then read the result:
 
 Use an include-enabled file with the standard Python logging module:
 
-.. python::
-
     import logging
     import includer
 
@@ -73,13 +64,11 @@ Use an include-enabled file with the standard Python logging module:
 
 """
 
-__docformat__ = "markdown"
-
-__all__ = ['Includer', 'IncludeError', 'preprocess']
-
 # ---------------------------------------------------------------------------
 # Imports
 # ---------------------------------------------------------------------------
+
+from __future__ import annotations # PEP 563 (allows annotation forward refs)
 
 import logging
 import os
@@ -87,17 +76,17 @@ import sys
 import re
 import tempfile
 import atexit
-import urllib.request, urllib.error, urllib.parse
-import urllib.parse
+import codecs
 
-import grizzled.exception
 from grizzled.file import unlink_quietly
 
-# ---------------------------------------------------------------------------
-# Exports
-# ---------------------------------------------------------------------------
+from typing import (Union, Sequence, AnyStr, TextIO, Optional, Iterable, List,
+                    Tuple, Any)
 
-__all__ = ['IncludeError', 'Includer', 'preprocess']
+
+__docformat__ = "markdown"
+
+__all__ = ['Includer', 'IncludeError', 'preprocess', 'MaxNestingExceededError']
 
 # ---------------------------------------------------------------------------
 # Logging
@@ -109,22 +98,32 @@ log = logging.getLogger('includer')
 # Public classes
 # ---------------------------------------------------------------------------
 
-class IncludeError(grizzled.exception.ExceptionWithMessage):
+class IncludeError(Exception):
     """
-    Thrown by ``Includer`` when an error occurs while processing the file.
-    An ``IncludeError`` object always contains a single string value that
+    Thrown by `Includer` when an error occurs while processing the file.
+    An `IncludeError` object always contains a single string value that
     contains an error message describing the problem.
     """
-    pass
+    def __init__(self, message):
+        Exception.__init__(self, message)
+        self.message = message
+
+
+class MaxNestingExceededError(IncludeError):
+    """
+    Thrown by `Includer` when the maximum include file nesting level is
+    exceeded.
+    """
+
 
 class Includer(object):
-    '''
-    An ``Includer`` object preprocesses a path or file-like object,
-    expanding include references. The resulting ``Includer`` object is a
+    """
+    An `Includer` object preprocesses a path or file-like object,
+    expanding include references. The resulting `Includer` object is a
     file-like object, offering the same methods and capabilities as an open
     file.
 
-    By default, ``Includer`` supports this include syntax::
+    By default, `Includer` supports this include syntax:
 
         %include "path"
         %include "url"
@@ -133,39 +132,41 @@ class Includer(object):
     expression, so it can be configured.
 
     See the module documentation for details.
-    '''
+    """
     def __init__(self,
-                 source,
-                 include_regex='^%include\s"([^"]+)"',
-                 max_nest_level=100,
-                 output=None):
+                 source: Union[TextIO, AnyStr],
+                 include_regex: AnyStr = '^%include\s"([^"]+)"',
+                 max_nest_level: int = 100,
+                 output: Optional[Union[TextIO, AnyStr]] = None,
+                 encoding: str = 'utf-8'):
         """
-        Create a new ``Includer`` object.
+        Create a new `Includer` object.
 
-        :Parameters:
-            source : file or str
-                The source to be read and expanded. May be an open file-like
-                object, a path name, or a URL string.
-            include_regex : str
-                Regular expression defining the include syntax. Must contain a
-                single parenthetical group that can be used to extract the
-                included file or URL.
-            max_nest_level : int
-                Maximum include nesting level. Exceeding this level will cause
-                ``Includer`` to throw an ``IncludeError``.
-            output : str or file
-                A string (path name) or file-like object to which to save the
-                expanded output.
+        **Parameters**
 
-        :raise IncludeError: On error
+        - `source` (`file` or `str`): The source to be read and expanded. May
+          be an open file-like object or a path name.
+        - `include_regex` (`str`):  Regular expression defining the include
+          syntax. Must contain a single parenthetical group that can be used
+          to extract the included file.
+        - `max_nest_level` (`int`): Maximum include nesting level. Exceeding
+          this level will cause `Includer` to throw an `IncludeError`.
+        - `output` (`str` or `file`): A string (path name) or file-like object
+          to which to save the expanded output.
+        - `encoding` (`str`): The encoding to use to open the files. Defaults
+          to "utf-8".
+
+        **Raises**
+
+        `IncludeError` on error
         """
 
+        self._encoding = encoding
         if isinstance(source, str):
-            f, is_url, name = self.__open(source, None, False)
+            f, name = self._open(source, None)
         else:
             # Assume file-like object.
             f = source
-            is_url = False
             try:
                 name = source.name
             except AttributeError:
@@ -173,259 +174,216 @@ class Includer(object):
 
         self.closed = False
         self.mode = None
-        self.__include_pattern = re.compile(include_regex)
-        self.__name = name
+        self._include_pattern = re.compile(include_regex)
+        self._name = name
 
         if output == None:
             from io import StringIO
             output = StringIO()
 
-        self.__maxnest = max_nest_level
-        self.__nested = 0
-        self.__process_includes(f, name, is_url, output)
-        self.__f = output
-        self.__f.seek(0)
+        self._maxnest = max_nest_level
+        self._nested = 0
+        self._process_includes(f, name, output)
+        self._f = output
+        self._f.seek(0)
 
     @property
-    def name(self):
+    def name(self) -> str:
         """
         Get the name of the file being processed.
         """
-        return self.__name
+        return self._name
 
-    def __iter__(self):
+    def __iter__(self) -> Iterable[line]:
         return self
 
-    def __next__(self):
-        """A file object is its own iterator.
-
-        :rtype: string
-        :return: the next line from the file
-
-        :raise StopIteration: end of file
-        :raise IncludeError: on error
-        """
+    def __next__(self) -> str:
+        """A file object is its own iterator."""
         line = self.readline()
         if (line == None) or (len(line) == 0):
             raise StopIteration
         return line
 
-    def close(self):
+    def close(self) -> None:
         """Close the includer, preventing any further I/O operations."""
         if not self.closed:
-            self.closed = true
-            self.__f.close()
-            del self.__f
+            self.closed = True
+            self._f.close()
+            del self._f
 
-    def fileno(self):
+    def fileno(self) -> int:
         """
         Get the file descriptor. Returns the descriptor of the file being
         read.
-
-        :rtype:  int
-        :return: the file descriptor of the file being read
         """
         _complain_if_closed(self.closed)
-        return self.__f.fileno()
+        return self._f.fileno()
 
-    def isatty(self):
+    def isatty(self) -> bool:
         """
         Determine whether the file being processed is a TTY or not.
-
-        :return: ``True`` or ``False``
         """
         _complain_if_closed(self.closed)
-        return self.__f.isatty()
+        return self._f.isatty()
 
-    def seek(self, pos, mode=0):
+    def seek(self, pos: int, mode: int = 0) -> None:
         """
         Seek to the specified file offset in the include-processed file.
 
-        :Parameters:
-            pos : int
-                file offset
-            mode : int
-                the seek mode, as specified to a Python file's ``seek()``
-                method
-        """
-        self.__f.seek(pos, mode)
+        **Parameters**
 
-    def tell(self):
+        - `pos` (`int`): file offset
+        - `mode` (`int`): the seek mode, as specified to a Python file's
+          `seek()` method
+        """
+        self._f.seek(pos, mode)
+
+    def tell(self) -> int:
         """
         Get the current file offset.
 
-        :rtype:  int
-        :return: current file offset
+        **Returns**
+
+        the current file offset
         """
         _complain_if_closed(self.closed)
-        return self.__f.tell()
+        return self._f.tell()
 
-    def read(self, n=-1):
+    def read(self, n: int = -1) -> Sequence[int]:
         """
         Read *n* bytes from the open file.
 
-        :Parameters:
-            n : int
-                Number of bytes to read. A negative number instructs
-                the method to read all remaining bytes.
+        **Parameters**
 
-        :return: the bytes read
+        - `n` (`int`): Number of bytes to read. A negative number instructs
+          the method to read all remaining bytes.
+
+        **Returns**
+
+        the bytes read
         """
         _complain_if_closed(self.closed)
-        return self.__f.read(n)
+        return self._f.read(n)
 
-    def readline(self, length=-1):
+    def readline(self, length: int = -1) -> str:
         """
         Read the next line from the file.
 
-        :Parameters:
-            length : int
-                a length hint, or negative if you don't care
+        **Parameters**
 
-        :rtype: str
-        :return: the line read
+        - `length` (`int`): a length hint, or negative if you don't care
+
+        **Returns**
+
+        the line read
         """
         _complain_if_closed(self.closed)
-        return self.__f.readline(length)
+        return self._f.readline(length)
 
-    def readlines(self, sizehint=0):
+    def readlines(self, sizehint: int = 0) -> List[str]:
         """
         Read all remaining lines in the file.
-
-        :rtype:  array
-        :return: array of lines
         """
         _complain_if_closed(self.closed)
-        return self.__f.readlines(sizehint)
+        return self._f.readlines(sizehint)
 
-    def truncate(self, size=None):
-        """Not supported, since ``Includer`` objects are read-only."""
+    def truncate(self, size: Optional[int] = None) -> None:
+        """Not supported, since `Includer` objects are read-only."""
         raise IncludeError('Includers are read-only file objects.')
 
-    def write(self, s):
-        """Not supported, since ``Includer`` objects are read-only."""
+    def write(self, s: str) -> None:
+        """Not supported, since `Includer` objects are read-only."""
         raise IncludeError('Includers are read-only file objects.')
 
-    def writelines(self, iterable):
-        """Not supported, since ``Includer`` objects are read-only."""
+    def writelines(self, iterable: Iterable[str]):
+        """Not supported, since `Includer` objects are read-only."""
         raise IncludeError('Includers are read-only file objects.')
 
-    def flush(self):
+    def flush(self) -> None:
         """No-op."""
         pass
 
-    def getvalue(self):
+    def getvalue(self) -> str:
         """
-        Retrieve the entire contents of the file, which includes expanded,
-        at any time before the ``close()`` method is called.
-
-        :rtype:  string
-        :return: a single string containing the contents of the file
+        Retrieve the entire contents of the file, as a string, with includes
+        expanded, at any time before the `close()` method is called.
         """
         return ''.join(self.readlines())
 
-    def __process_includes(self, file_in, filename, is_url, file_out):
-        log.debug('Processing includes in "%s", is_url=%s' % (filename, is_url))
+    def _process_includes(self,
+                          file_in: TextIO,
+                          filename: str,
+                          file_out: TextIO) -> None:
+        log.debug(f'Processing includes in "{filename}"')
 
         for line in file_in:
-            match = self.__include_pattern.search(line)
+            match = self._include_pattern.search(line)
             if match:
-                if self.__nested >= self.__maxnest:
-                    raise IncludeError(
-                        'Exceeded maximum include recursion depth of {0}'.format(
-                            self.__maxnest
-                        )
+                if self._nested >= self._maxnest:
+                    raise MaxNestingExceededError(
+                        f'Exceeded maximum include depth of {self._maxnest}'
                     )
 
                 inc_name = match.group(1)
-                logging.debug('Found include directive: %s' % line[:-1])
-                f, included_is_url, included_name = self.__open(inc_name,
-                                                                filename,
-                                                                is_url)
-                self.__nested += 1
-                self.__process_includes(f, filename, is_url, file_out)
-                self.__nested -= 1
+                log.debug(f'Found include directive: {line[:-1]}')
+                f, included_name = self._open(inc_name, filename)
+                self._nested += 1
+                self._process_includes(f, filename, file_out)
+                self._nested -= 1
             else:
                 file_out.write(line)
 
-    def __open(self, name_to_open, enclosing_file, enclosing_file_is_url):
-        is_url = False
-        openFunc = None
+    def _open(self,
+              name_to_open: str,
+              enclosing_file: Optional[str]) -> Tuple[TextIO, str]:
 
-        parsed_url = urllib.parse.urlparse(name_to_open)
-
-        # Account for Windows drive letters.
-
-        if (parsed_url.scheme != '') and (len(parsed_url.scheme) > 1):
-            openFunc = urllib.request.urlopen
-            is_url = True
-
-        else:
-            # It's not a URL. What we do now depends on the including file.
-
-            if enclosing_file_is_url:
-                # Use the parent URL as the base URL.
-
-                name_to_open = urllib.parse.urljoin(enclosing_file, name_to_open)
-                open_func = urllib.request.urlopen
-                is_url = True
-
-            elif not os.path.isabs(name_to_open):
-                # Not an absolute file. Base it on the parent.
-
-                enclosing_dir = None
-                if enclosing_file == None:
-                    enclosing_dir = os.getcwd()
-                else:
-                    enclosing_dir = os.path.dirname(enclosing_file)
-
-                name_to_open = os.path.join(enclosing_dir, name_to_open)
-                open_func = open
-
+        if not os.path.isabs(name_to_open):
+            # Not an absolute file. Base it on the parent.
+            if enclosing_file == None:
+                enclosing_dir = os.getcwd()
             else:
-                open_func = open
+                enclosing_dir = os.path.dirname(enclosing_file)
 
-        assert(name_to_open != None)
-        assert(open_func != None)
+            name_to_open = os.path.join(enclosing_dir, name_to_open)
 
         try:
-            log.debug('Opening "%s"' % name_to_open)
-            f = open_func(name_to_open)
+            log.debug(f'Opening "{name_to_open}" with encoding {self._encoding}')
+            f = codecs.open(name_to_open, mode='r', encoding=self._encoding)
         except:
             raise IncludeError(
-                'Unable to open "{0}" as a file or a URL'.format(name_to_open)
+                f'Unable to open "{name_to_open}".'
             )
-        return (f, is_url, name_to_open)
+        return (f, name_to_open)
 
 # ---------------------------------------------------------------------------
 # Public functions
 # ---------------------------------------------------------------------------
 
-def preprocess(file_or_url,
-               output=None,
-               temp_suffix='.txt',
-               temp_prefix='inc'):
+def preprocess(file: Union[TextIO, str],
+               encoding: str = 'utf8',
+               output: Optional[TextIO] = None,
+               temp_suffix: str = '.txt',
+               temp_prefix: str = 'inc'):
     """
     Process all include directives in the specified file, returning a path
     to a temporary file that contains the results of the expansion. The
     temporary file is automatically removed when the program exits, though
     the caller is free to remove it whenever it is no longer needed.
 
-    :Parameters:
-        file_or_url : file or str
-            URL or path to file to be expanded; or, a file-like object
-        encoding : str
-            String encoding for input file/URL. Defaults to UTF-8.
-        output : file
-            A file or file-like object to receive the output.
-        temp_suffix : str
-            suffix to use with temporary file that holds preprocessed output
-        temp_prefix : str
-            prefix to use with temporary file that holds preprocessed output
+    **Parameters**
 
-    :rtype:  string
-    :return: ``output``, if ``output`` is not ``None``; otherwise, the path to
-             temporary file containing expanded content
+    - `file` (`file` or `str`): path to file to be expanded, or file-like object
+    - `encoding` (`str`): String encoding for input file. Defaults to UTF-8.
+    - `output` (`file`): A file or file-like object to receive the output.
+    - `temp_suffix` (`str`): suffix to use with temporary file that holds
+      preprocessed output
+    - `temp_prefix` (`str`): prefix to use with temporary file that holds
+      preprocessed output
+
+    **Returns**
+
+    `output`, if `output` is not `None`; otherwise, the path to temporary file
+    containing expanded content
     """
     result = None
     path = None
@@ -438,7 +396,7 @@ def preprocess(file_or_url,
     else:
         result = output
 
-    Includer(file_or_url, output=output)
+    Includer(file, output=output, encoding=encoding)
     return result
 
 
@@ -446,7 +404,7 @@ def preprocess(file_or_url,
 # Private functions
 # ---------------------------------------------------------------------------
 
-def _complain_if_closed(closed):
+def _complain_if_closed(closed: bool) -> None:
     if closed:
         raise IncludeError("I/O operation on closed file")